--story=872097271 rds 注册触发事件规则

+++ src/backend/ci/ext/tencent/rds/biz-rds-tencent/src/main/kotlin/com/tencent/devops/rds/utils/RdsPipelineUtils.kt
@@ -29,7 +29,7 @@ package com.tencent.devops.rds.utils
 
 object RdsPipelineUtils {
 
-    fun genBKProjectCode(productId: Int) = "RDS_$productId"
+    fun genBKProjectCode(productId: Long) = "RDS_$productId"
 
-    fun genBKPipelineName(productId: Int) = "RDS_" + productId + "_" + System.currentTimeMillis()
+    fun genBKPipelineName(productId: Long) = "RDS_" + productId + "_" + System.currentTimeMillis()
 }
